Resolve strange elm bug by removing unnecessary event emission
Full stacktrace:

Uncaught TypeError: stepper is not a function
    at sendToApp (Main.elm:1885)
    at HTMLElement.callback (Main.elm:3001)
    at HTMLElement.connectedCallback (codemirror.js:144)
    at _VirtualDom_applyPatchRedraw (Main.elm:3798)
    at _VirtualDom_applyPatch (Main.elm:3716)
    at _VirtualDom_applyPatchesHelp (Main.elm:3702)
    at _VirtualDom_applyPatches (Main.elm:3693)
    at Main.elm:3948
    at _Browser_makeAnimator (Main.elm:4009)
    at Main.elm:3944

In the elm core:

function sendToApp(msg, viewMetadata)
{
        var pair = A2(update, msg, model);
        stepper(model = pair.a, viewMetadata); // error happening here
        _Platform_enqueueEffects(managers, pair.b, subscriptions(model));
}

Looks like this event emission from Codemirror was happening earlier
than Elm could handle. There is no record of this error on the internet,
so I should report this to the Elm team someday.