Implement tutorial mode
Improve the main Elm app to allow for a special "tutorial mode" in which the opponent select acts as a "chapter" select instead. Add a new pane on the left to display tutorial text when this is the case. For now, load the tutorial data from a url passed as a query parameter.